corrigi um erro que tinha na hora de subtrair o total, estava em inteiro, entao mudei para parseFloat

+++ versao2.0/script/etapa1.js
@@ -269,7 +269,7 @@ function removerPrato(e){
 }
 function calculaTotalQuandoRemove(json,idPrato,objQuantPratoFeitos){
 	//subtrai o valor do prato excluido no total 
-	let valorTotal = parseInt(sessionStorage.getItem("valorTotal"));
+	let valorTotal = parseFloat(sessionStorage.getItem("valorTotal"));
 	let preco = json[parseInt(idPrato)].preco;
 	
 	let quant = objQuantPratoFeitos;
